requirements.txt
google-auth==2.25.2
Pillow==10.1.0
google-cloud-vision==3.5.0
